Updated alignem.py and alignem_swift.py to work together on scale_1
This version is capable of aligning a single stack at full scale using
pyswift_tui.py from start to finish.

This version was developed in the "JSON_Centric" branch of the
repository, but since its functionality is approximately that of the
version in the development branch, this branch will be merged into the
development branch shortly. Furthermore, this version currently has its
own copy of pyswift_tui.py so it should not interfere with the version
in the "python" directory.

To align with this version:

  1) python alignem_swift.py
  2) Images / Empty into / ref   (leaves a space in the reference)
  3) Images / Import into / ref  (import all but the last of stack)
  4) Images / Import into / base (import the full stack - all)
  5) File / Set Destination  (a place for output images and copies)
  6) Click "GenScales" button  (actually makes links in "scale_1")
  7) Click "Align All SWiFT"  (runs the alignment and loads results)

The project may also be saved with File / Save Project As.
The project can be reloaded with File / Open Project.