Added functions to help manage scales, scale strings, and scale keys
Scale keys in the JSON file are now prefixed with "scale_" to make them
easier to spot. This means that scales can now exist as key values
(with the "scale_" prefix), scale strings (for display), and scale
values for computation.

This commit adds two functions to convert between scale values and
scale keys:

  get_scale_val()  # Accepts a scale key ("scale_#") or scale string
  get_scale_key()  # Accepts either an int or an int in a string

These functions are used in a number of places to ensure that the proper
representation of a scale is used in the proper context. This also keeps
all of the hard-coded "scale_" strings in one location.