Create a list of keys for iteration in remove_all_images.
This code works in Python 2 but fails in Python 3:

  def remove_all_images ( self ):
      keys = self.images.keys()
      for k in keys:
          self.remove_image_reference ( k )
      self.images = {}

This is the error:

  Traceback (most recent call last):
    File "alignem_swift.py", line 223, in remove_aligned
      alignem.image_library.remove_all_images()
    File
  ".../source/PySide2/alignem.py",
  line 162, in remove_all_images
      for k in keys:
  RuntimeError: dictionary changed size during iteration

It appears that Python3 {}.keys() function creates an "iterator"
rather than a list. Changing the dictionary in the loop appears
to break the iterator (thus the error). On recommended way around
this is to simply create a list from the keys iterator:

  def remove_all_images ( self ):
      keys = list(self.images.keys())
      for k in keys:
          self.remove_image_reference ( k )
      self.images = {}

That's what was implemented in this commit.