Set title bar to project file, and DON'T USE "os.realpath"!!
The "os.realpath" function resolves symbolic links to give the actual
file name (not the linked name). This was creating potential problems
when we use symbolic links to the original images in "scale_1". In
particular, the JSON files were being written with the original file
names rather than the names within the "scale_1" directory.

This commit adds a "pathkey" function to the ImageLibrary to consolidate
the path usage within that library. The new "pathkey" function uses
"os.abspath" instead of "os.realpath". This should make any such future
changes a bit easier.