Force use of scale_1 when generating scales.
The previous version continued to use the originally loaded file names
for the "scale_1" files after scaling the other images.

This version uses the "Generate Scales" button to not only make the
"scale_1" directory with links to originals, but also changes the names
in the stack to use the new "scale_1" paths (which should be links to
the originals). Note that this required a special check to be sure that
the original source images were not already the "scale_1" images when
regenerating the scales. The previous version simply unlinked and
relinked the "scale_1" files since they were being linked to other files
outside of the local project tree. But now that the "scale_1" files may
already be in the project tree, they cannot be unlinked without first
checking that they are not now also the "original" images.