Copy images when linking fails (Winodws 10 users may sometimes not link)
Windows 10 does not allow normal users to create symbolic links by
default. This commit checks for failure of the link and follows up with
a shutil.copy() call as needed. This may result in a duplication of some
very large image files, so a notification is printed when that happens.
It would be better if users had the appropriate permissions to make
links, but this commit at least provides a backup if they cannot.