Disabled showing of full paths by default (generally too long)
Note that the threaded loading sometimes loads the same image multiple
times. This may be due to an uncertainty about whether an image has been
scheduled for loading or not.

Right now, the current check is whether it is actually loaded before
scheduling it to be loaded. However, a load may be in progress from an
earlier request. This may cause images to be loaded multiple times. An
additional state of "loading in progress" may be needed to fix this.