Added centering of images (default when opening a project)
This commit enables both centering and showing actual size via the menu.

The "centering" is biased downward on the screen to leave more room for
the annotations (file name etc) at the top of the image.

The images are zoomed by increments of the mouse wheel until they fit
within the window. This ensures that the full scale of 1.0 can be found
while zooming. Otherwise, it might not be possible to find the original
size among the mouse wheel choices. For this reason, the centered image
does not always fill the window completely.