Added mouse click connection between alignem and alignem_swift.
This commit demonstrates the handling of mouse clicks inside of
alignem_swift.py. The mouse move and mouse down callbacks are registered
in the main, and then they are called whenever they happen. They decide
for themselves (based on anything ... especially user settings) what to
do and then do it.

Both the raw screen positions and the position in the image are returned
along with the "role" of the image clicked.