Added code to support match points (not fully working yet).
This version supports adding match points to the metadata. The match
points are added to the metadata as raw match points (list of pairs),
and also as a series of graphics commands ( "circle(x,y,r)" ). The
graphics commands are just strings, so they can take any form that we
define (circle, point, line, polyline, etc). This does mean that the
annotations have to be "parsed" every time they're drawn, but since we
have so little metadata per image, this might be fine.

For example, this shows 5 match points in the JSON file:

   "metadata": {
     "match_points": [
       [ 118.02, 158.394 ],
       [ 0.8999, 2.87399 ],
       [ 5.37999, 338.234 ],
       [ 341.38, 338.8739 ],
       [ 341.38,  -0.325994 ]
     ],
     "annotations": [
       "circle(118.02,158.394,10)",
       "circle(0.90,2.874,10)",
       "circle(5.38,338.234,10)",
       "circle(341.38,338.874,10)",
       "circle(341.38,-0.326,10)"
     ]
   }

There is some redundancy there, but the separation allows each
representation to better fit its purpose.

This is just a rough prototype and is open to revision.