Computed proper locations for match points.
The prototype version from last night did not properly display the
locations of the match points. It also showed match points as squares.
This version computes locations correctly and shows match points as
circles.

This version also moved the "Show Border" to the "Show" menu, and added
support for a "square" annotation command.