Added color to match point annotations and added ability to clear.
This commit adds simple colors (R,G,B,Y,M,C) to the match points. This
commit also adds the ability to clear all the match points on a given
layer.

As protection against accidentally deleting the match points, the
"Match" mode must be selected in order for the "Clear" button to work.