Removed colors from JSON, but they are still recognized.
The colors added too much extra text to the JSON as they had been
implemented, so they were removed by simply not writing them. In that
case, the default colors will be used:

  0 --> Red
  1 --> Green
  2 --> Blue
  3 --> Yellow
  4 --> Magenta
  5 --> Cyan

Note that colors in the JSON are still recognized if they exist, and
alignem.py will honor them. But alignem_swift.py no longer generates
them, and alignem.py now uses the defaults when not colors are given.