Commented out ShowMessage calls to find "Segmentation fault: 11"
Under certain conditions, the progress bar appears to cause a
segmentation fault. It's not clear why that is happening since
the code uses the somewhat opaque QThread class for loading. The
final messages (with debug set to 101) are:

  Total images to scale = 42
  GenScalesThread.run inside alignem_swift called
  Create images at all scales: [1, 2, 3, 4, 5, 6]
  Creating images for scale 1
  QPainter::begin: A paint device can only be painted by one painter at a time.
  QPainter::setCompositionMode: Painter not active
  Segmentation fault: 11

The current class ("GenScalesThread") is a subclass of QThread,
and it had been calling the "showMessage()" function of the status
bar. These calls were commented out in this version in case that's
an improper use of data across the thread boundary. While that
seemed (initially) to improve the situation, it eventually failed
with the messages listed above.

At this time, the threaded version for generating scales is not
working, so the normal version should continue to be used.