Added an "Import Base Images" menu option
The previously existing menu options allowed for importing images
into any role (including "base"). That provides the most flexibility
but can also lead to more confusion. This new menu option just gives
a simpler way to do what was already possible before. This was done
in anticipation of a release to the alignment team.