Added a "Link All Stacks" button to link all stacks for all scales.
This is another convenience function added to make the tool easier
to use. The "Link All Stacks" button is currently in the second row
but will likely be moved to the first row. There may still be times
where the linking of a single stack is desired, so the "Link Stack"
button may be preserved somewhere else in the interface.

Note that this version duplicates the "skip" flag for each scale
based on the skip values of the original ("base") stack.