Added "Set Project Destination" and "Set Other Destination..." options
The previous "Set Destination..." menu item allowed the setting of any
destination within the file system. That is still supported in the now
renamed "Set Other Destination..." menu item.

Additionally (and intended as the default mode of operation), the menu
also contains a "Set Project Destination" option. This option sets the
destination directory to be the same as the project file name without
the ".json" extension. The directory is created if it doesn't exist.
This follows the pattern used by CellBlender projects.

In the future, the creation of the destination directory might be done
whenever the project file is saved. This current version was written
to be more explicit.