Began working on scale_change_callback.
As mentioned in the previous commit, values are swapped properly
between the data model and the widgets when scrolling through the
layers (sections), but not when changing scales. So the values in
the widgets will still be from the previous scale when scales are
changed. These values will go into the data model when layers are
changed at that scale, and then subsequent values in that scale
will be correct. But the values at the scale change will be wrong.

The work around is to always check and modify values as needed when
changing scales. If there is a layer where the values are the same
across scales, then that would be the best place to change scales.
Otherwise, change scale, check values at that layer, fix values at
that layer, and all other layers will be fine within that scale.

This commit is a start at fixing this problem.