Function/Variable name changes from "data_change" to "view_change"
The original plan for the API was to provide a callback whenever a value
(data) is changed. However, that hasn't been implemented in that form.
Instead, the client application is notified when the layer is changed
(and will eventually be notified when the scale is changed). This is
much more similar to a "frame change" mechanism than a data changed
callback. This commit begins to change the names to reflect what is
currently being done so that it might be generalized to include scale
changes as well as layer changes. This is more accurately reflected as
a generalized "view" change, and the names in this commit have been
changed appropriately.

This commit also adds "prev_scale_key" and "next_scale_key" parameters
to the newly renamed "view_change_callback" function to handle scale
changes as well as layer changes.