"view_change_callback" now working between generic layers (from any scale)
This change uses the prev and next scale keys along with the prev and next
layer numbers to choose between the layers (possibly across scales) to swap
values through the GUI widgets. While the "view_change_callback" now handles
both layer and scale changes, it is not yet being called when scales change.