Use the view_change_callback with new_data_model=True to load project.
The view_change_callback function has a flag that allows it to ignore
the current values in the widget fields and simply over-write them from
the data model. This commit uses that function to force the proper data
field values when loading a new project.

Note that a "double free" error has been appearing similar to:

  Python(1400,0x700001e59000) malloc: *** error for object 0x7fdf1ed031d0: double free
  *** set a breakpoint in malloc_error_break to debug
    load_image_worker finished for: ".../proj/scale_5/img_aligned/vj_4k4k_2.jpg"
  Abort trap: 6

It's not clear why a "double free" should be happening from Python.
This does not happen every time, but it's happened enough to make a note.