Remove the scale_change_callback and replace with view_change_callback
The "view_change_callback" now handles changes to both the layer and scale.
It is intended to reflect a user request to change the "view" where the
"view" includes both layer and scale and possibly other keys.

This commit removes the "scale_change_callback" and replaces it with an
appropriate call to view_change_callback.