Added "plumbing" to get Align Forward information into pyswift_tui.
The quick plan to support "Align Forward" was to rebuild the data model
outside of PySwift_TUI. However, after a few attempts, it was not clear
how to best implement that functionality. This commit just contains the
"plumbing" to pass the user's selected starting layer and the number of
layers to be aligned into run_json_project (as start_layer, num_layers).
These are passed as 0 and -1 respectively.

This approach will need some more thought before implementing.