New approach for Align Forward by skipping operations (still not done)
This commit begins to implement the idea mentioned in the last commit
message. It allows all of the normal alignment logic to proceed as if
it were aligning the full stack. But it attempts to circumvent the more
compute intensive operations for those layers that are not requested to
be aligned in this pass. The bypassed operations can be found in the code
by looking for lines such as:

  if i in range_to_process:
    # Do something that's not needed for aligned or excluded layers

The "range_to_process" is a list of those layer numbers that should be
processed for alignment in the current run.

This logic works to some degree, except the Cumulative Affine Matrices
are not being handled properly. So the images do not align properly when
aligning forward from any non-zero starting position.

Note that aligning the entire stack or any subset from the beginning still
works properly as it did before.