Extensive change from "print" to "print_debug" for better output control.
This change converted almost all of the normal "print" statements to
"print_debug" statements so the output can be limited. This was done to
eventually help debug the subtle problems in the "Align Forward" code.

Note that without print statements, the interface may sometimes respond
much faster than with print statements. This may reveal latent problems
like this one:

  Python(2622,0x7fff7d4e8000) malloc: *** error for object 0x7fed83e4c418:
    incorrect checksum for freed object - object was probably modified after being freed.
  *** set a breakpoint in malloc_error_break to debug
  Abort trap: 6

This error was noted on the Macintosh when scrolling rapidly through
the stack. It immediately terminated the program.