Added finer debug levels to print AFM and CAFM only for Align Forward
Most normal debug statements range from 10 to 100. So to print out just
the AFM/CAFM data, it would be helpful to have a level with higher
priority than 10 (lower numbers). This commit added levels 1-9 to allow
that finer grained output, and also prints out the AFM/CFM as it aligns.