BAD VERSION: This version contains a hard-coded "rect" in saveAligned
The process of debugging (lots of "print_debug" statments) revealed
that the "rect" calculated by "BoundingRect" required the whole stack.
However, when aligning forward, the whole stack is not available, and
"BoundingRect" produces different results as each section of the stack
is processed.

This version preserves the test of setting the "rect" equal to the
expected values of:

   [-116, -116, 914, 914]

Those were the values produced when the entire test stack was aligned,
and they produced a good result when aligning only parts of the stack.

The likely solution will be to tag each item in the "align_list" so
that it can be used as needed (for things like computing the bounding
rect) and skipped by the "expensive" operations like running swim.

The align_list is currently a list of 2-item lists where the first
item is the layer index and the second item is the alignment process.
To "tag" each layer, a third item will likely be added to flag whether
the layer is really intended to be aligned or not. It is anticipated
that the stuffing of 3 unnamed items in a list is prone to errors, so
the 2-item lists will likely be turned into 3-item dictionaries.