BAD VERSION: Converted align_list from a list of lists to a list of dictionaries
This version also uses a fixed bounding rectangle - DON'T USE!

The align_list had contained a list of items where each item was:

  [i, process]

The items were referenced somewhat cryptically by index:

  align_idx = item[0]
  align_item = item[1]

This change simply replaces each list in the list with a dictionary.

  {'i':i, 'proc':process, 'do':flag}

Then the previous code becomes:

  align_idx = item['i']
  align_item = item['proc']

The additional "do" flag will be used to determine which items
in the list should perform the actual work of alignment. While
the flag is stored, it is not currently being used. That will
hopefully be completed in the next commit.