Store bias data in bias_data_path for a given scale
Also Get int values of scale keys from data model in a way that is compatible with
the old keys (e.g. '1') and new keys (e.g. 'scale_1')

And concatenate 'scale_' with int scale value when accessing the scale dictionary
in the data model.  Bob had fixed this in one place but there were 3 other places
where this needed to be done in alignem_swift.py