Add function evaluate_project_status()
This function is created to refactor run_json_project into reuseable pieces

The returns a small dictionary with status information pertaining to the
alignment completion status for the whole project and each scale.

This information is need to determine what steps are required and valid in
continuing the alignement process for each scale in the project.

The refactoring makes the scale_done, and scale_tbd args to run_json_project
unnecessary so these are removed.