Changes to use the {"method:data":{"alignment_option":...}} from the scale
Previous versions had been storing an "alignment_option" at the layer
level for every layer. The previous commit added an "alignment_option" at
the scale level but didn't use it. This commit adds support for that option
by using it. The alignment_option at the layer level remains in the code,
but it should not be used and will be phased out after testing.