Added "ensure_proper_data_structure" function to alignem_swift.py
The "alignem.py" program handles the images and the general stack.
Additional data can be added to the general stack by alignment
algorithms. They are responsible for managing that data and
ensuring its integrity.

This change to alignem_swift.py adds an "ensure_proper_data_struture"
function to do that prior to calling the actual alignment code. The
current version uses "hard coded" defaults (the same ones used in
the FloatFields of those values). However, it should strive to get
them from the current settings in a future version.