Removed the SNR display after removing aligned images.
The "remove_aligned" function had only been removing the aligned images
and not the associated data from the data model. This was most noticable
with the SNR value which would still be displayed from the old alignment.

A first attempt at fixing this (still in the code as a comment) was to
simply remove (dict.pop) the entire "method_results" item. However, this
caused problems in line 286 of pyswift_tui.py which expected to have a
method_results section. So rather than removing the entire methos results,
this change sets both the "affine_matrix" and "snr" values to None.