Removed import of "qApp" from PySide2.QtCore
The version of PySide2 installed on TACC failed to import qApp from
the PySide2.QtCore module. Fortunately, qApp is included but not used,
so it was removed from the list of imports in order to get it to run
on TACC. This commit reflects the change made during the meeting on
the TACC installation.