Added code to print up to 4 parameters in all Python versions.
The "print_debug" function was intended to be a lightweight and easy
to use mechanism for controlling the amount of output. It was also
designed to take an arbitrary number of parameters. However, it used
a variable argument mechanism not available in Python2.

This commit replaces that function with one that takes up to 4 optional
parameters (defaulted to None). It determines what should be printed
based on the number of non-None arguments. This isn't optimal, but it
works well enough for now.