Added support for Revision and Source Hash to alignem_swift.py
This version displays the revision (if known) and the source hash on
a third row of the control panel. There is probably a better place to
display this information, but for now this makes it very obvious.

The new revision system is based on a self-hash of the code when the
program is first run. Since the files are so small this should be a
trivial task. So the displayed value should detect any changes that
have been made (intentionally or accidentally).

The program always checks its current version against the value stored
in source_info.json (assumes that it was run from the source directory
for now). If the version doesn't match, then the new version is written
to source_info.json. This automatically updates that file so there is
no manual intervention required by developers. As long as they have at
least run their code before pushing, the value in the JSON file should
match the value that the code gets from the hash of all source files.

All developers should ALWAYS run their code before pushing!!

The source_info.json file also contains a section for "tagged_versions".
This is a dictionary of revision tags keyed by the hash of the source.
If the source hash is found in that dictionary, then it currently returns
a string containing the revision number. These should be added whenever
a tagged version is committed. Note that the hash value is the hash of
the source files and NOT the GIT hash. The git hash can be found by
searching the repository for any versions where the source_info.json
file has a "current_hash" with that value.