Added support for detecting/removing GIT Merge Markers from source_info.json
When GIT detects a change to source_info.json, it will add merge
markers indicating the change:

  {
  <<<<<<< HEAD
    "current_hash": "052f4245058a8f4f84886d27cb1c096f7ca2e483",
  =======
    "current_hash": "03f477aebda2d6656ae6ea911fa5ee98278fdd72",
  >>>>>>> d8df7c5d9d97a774787d8ca4d3d089ce73ed0098
    "tagged_versions": {
      "150e0acd86a3c57a7d79fde4b3c4e225b55f26e4": "0.2.3",
      "2eac19472e2631173bbbe09d3c4251e219508602": "0.2.2",
      "5d9fc94902a283bf458d7be61f6dccaafd5fb553": "0.2.1"
    }
  }

This makes the entire file non-legal-JSON and it cannot be loaded
with the JSON library. The previous solution required manually
editing the file to remove the markers and all text in between.
Then running alignem again would add the "current_hash" value
for the current source code.

Since this manual editing would need to be done for every merge,
this commit adds a small amount of code to source_tracker.py to
do that automatically. Since source_tracker.py is already reading
and updating the file, it is the logical place to detect merge
markers and remove them.

Of course, any other merge problems will need to be fixed with
the normal process.