Added a detailed description regarding merging and source_info.json
The "source_info.json" file is automatically updated every time any
of the "AlignEM" source files change. However, the update is a whole
new source file hash, and GIT has no way to merge two of these when
they collide. So when merging any source code modifications, this
file would normally need to be updated. The previous commit adds
code to handle this automatically whenever either of the "AlignEM"
programs are run. This current commit adds text to the README.md
file (in docs/development) to explain the process.