Initial import of get_image_size.py
This module is a fast and efficient way to get image width and height
from many different image file formats, using only pure python and core
python modules.  We're not yet using this method in the alignment code.

TODO:
We will look in the alignment code for places where we just need to find
out the image size to setup the alignment process where there is no need
to read-in the whole image. In those cases we can save significant IO
overhead by using get_image_size.py.