Only copy method results from coarser to finer scales when climbing.
The previous version copied the entire stack from a coarser scale to
the final scale. This had the effect of forcing coarser images into the
stack of finer images.

This commit does not attempt to copy the entire stack. Instead, it just
copies the method_results portion of the dictionary. This seems to work
for some "align forward" examples, but the code should probably be
rewritten to be much cleaner.