task_queue.py now makes actual processes only at work_q activation time
Previously add_task would create an actual process at the time the work
was put in the work_q.  This process would then sit and wait for input
on its stdin pipe.  This method can use system process table resources
and which can be exhausted with large numbers of tasks in the work_q.

With this commit add_task now adds a process_template to the work_q.
The process_template is just a string containing the Popen command.
The process_template is eval'd to create the actual process only when
a thread worker pulls it from the work_q.

Also, it is important to close the stdin, stdout, and stderr pipes after the
process completes.  If these pipes are left open they continue to consume
system resources for the duration of alignem_swift.py session.