Added centering based on all images loaded in a given role.
This commit adds an additional option to the centering routines named
"all_images_in_stack". This option is now defaulted to "True". When
True, this routine will find the maximum width and height of all
images THAT ARE CURRENTLY LOADED in that role in the stack. The routine
only searches those that are loaded to keep from forcing a load of all
the images in the entire project just to get their sizes.