Added "DontUseNativeDialog" flag to File Dialogs.
Some PySide2 Dialogs have an optional capability to draw a "Qt" style
dialog rather than the "native" style dialog. This was experimented
with in earlier versions of the code (see the difference for this
commit). It was decided that the native dialogs would be more natural,
so this option was commented out. However, recent running on some new
platforms has revealed shortcomings in the native dialogs. So this
commit sets the "DontUseNativeDialog" flag for all of the QFileDialog
boxes that select files.