Added a result of "None" when no valid alignment method is found
The "align" method in alignment_process withing align_swiftir.py used
an "if" structure that required one of two matching states to be met
before assigning a result. When neither condition was met, the result
was undefined and failed with the following error:

  Traceback (most recent call last):
    File "pyswift_tui.py", line 885, in <module>
      num_layers=num_layers )
    File "pyswift_tui.py", line 669, in run_json_project
      c_afm = align_item.align(c_afm,save=False)
    File "/Users/bob/proj/swiftir/source/PySide2/align_swiftir.py", line 148, in align
      return result
  UnboundLocalError: local variable 'result' referenced before assignment

This commit sets the result to "None" prior to the "if" statements to
avoid this error and return a "None" result.