Added code to automatically set the "selected_method" to "Auto Swim Align"
The previous code was failing when run from the command line with an
unaligned data model because the ['align_to_ref_method']['selected_method']
was not set.

This commit sets that value to "Auto Swim Align" if not otherwise set. This
is done in pyswift_tui.py inside "run_json_project".