Added "alone" flag to run_json_project as needed for parallel use
The "alone" flag is intended to signal alignment in isolation of any
of the other layers. This will be needed for parallel use. It disables
use of data from other layers.

Note that there might still be a problem because some of the alignment
code appears to treat layer 0 in a special way. This dependency on layer
0 should instead depend on whether there is a reference image for that
layer. When aligning a single layer, there will only be one layer (0),
but that layer will have a reference image. So the "copy" mode should
depend on whether there is a reference or not, and NOT on the layer
index. This would be a more robust operation anyway.