Initial import of image_apply_affine.py
This is a commandline and importable function tool for applying an affine tmatrix
to an input image file (read from the file system) and writing the output image
to the file system.  This will be used for future versions of the code when
we want to apply the Cafm to the base image to a create the aligned image.
The command line interface allows this method to be run concurrently via
the task queue.

Not yet connected to the rest of the UI.