Create the temp file in the project subdirectory.
This version still uses the tempfile.NamedTemporaryFile() function to
ensure that the file name doesn't conflict with anything else. This was
originally done with the thought of running multiple alignments at the
same time. However, for other reasons it may not be practical to have
multiple alignments running at the same time. In that case, any file
name that is known to not conflict with other names would be fine.