Added single_alignment_job.py to call run_json_project as a task
This version of alignem_swift.py uses a temporary command line argument:

  -p [ 0 | 1 ]

to indicate aligning in parallel (or not). When non-zero, this causes
project_runner.py to attempt to run the alignments via a task queue.

At this time, it doesn't seem to work because the tasks are started
with a working directory of "None" which fails.

Note that the "-p" parameter defaults to 0 which causes project_runner
to call pyswift_tui.run_json_project directly. This works, so this
version still functions as before when run with default parameters.