Initial refactoring of the alignment process pipe line
align_swiftir.py now operates directly on a single layer_dict item from a datamodel
It calculates only the afm and alignement stats and stores these directly in the method_results
dictionary of the layer_dict item.

Several in-line sections of run_json_project have been refactored into separate function calls:

1) SetStackCafm() computes the c_afm with optional bias correction.
   This is an inherently serial process but is fast and efficient to compute.

2) save_bias_analysis() analyzes the final bias stats and saves these in the destination_dir
   Again, this is an inherently serial process but is fast and efficient to compute.

The final step of applying the c_afms and saving the aligned images in the destination_dir
is still done in-line at the end of run_json_project.  This should be refactored into
a separate function that makes use of image_apply_affine.py, optionally in parallel via
the TaskQueue.