Added a command line option to disable/enable C mode.
The new "-c" option takes an integer value (0 or not) that specifies
whether to use the "C" code or not. This is handy for running in an
environment where one version or the other isn't working and the
command line is used to start in the program in the desired mode.