Initial rubber band experimentation (always draws rubber band).
This version always draws with a rubber band when doing any click/drag
in any of the windows. This version was adapted from code at:

  https://doc.qt.io/qtforpython/PySide2/QtWidgets/QRubberBand.html

Note that the code in that example appears to be adapted from C++ code
because it does not consistently use the "self" with class variables
that should require it. For example, "origin" is set in the mouse
pressed callback, but it is not preserved in the "self". It might
be global, but that's not shown. Even worse, the "self.rubberBand"
is set in the mouse pressed callback but it's used without "self"
in both the mouseMoveEvent and mouseReleaseEvent.