Changed output from temp file to "project_runner_job_file.json"
The previous temp file creation was replaced with the creation of a
fixed named file ("project_runner_job_file.json") in the project's
directory (along with the "scale_#" directories). This was done to
try to address a parallel running problem experienced on TACC.

This commit also commented out many of the print statements that
might also be a source of I/O contention. But this was not an
exhaustive effort.