Change to avoid use of 'eval' when create the processes in task_queue.py
Attempting to fix problem with running tasks in the task_queue seeming
to block each other and slow each other down.  Still a problem...

Also save temporary json file with a fixed file name as an experiment to see
if the unique id temp file was part of the problem with parallel processes
seeming to block each other.  This change did not help.