Changes made in an attempt to identify the process blocking in the task_queue
processes seem to block each other in the task_queue, perhaps due aggregation of stdout
from each process into the main thread of the task_queue. Perhaps Python's GIL is
the culprit.  In this commit I've added a new version of task_queue.py called task_queue2.py
which uses the same API as task_queue but which greatly simplifies the infrastructure
under the hood.  task_queue2 no longer uses the task_wrapper and OutputQueue as abstraction
layers in process creation. Instead the worker threads use Popen and proc.communicate directly
for their work item.  Still, the stdout and stderr are collected by the worker threads and
aggregated back into the task_dict for their work item -- and this is perhaps where the GIL
contention is killing us.

Also, made changes to reduce the IO to stdout and stderr for the various jobs we run through
the task_queue.

In the end, benchmarks reveal that parallel gen_scales via iscale2 can generate images for
scale_4 of r34 (194 images 4k x 4k) in 20 seconds (4 cpus 3.1Ghz core i7, MacOSX) and 65 seconds
on 1 cpu.  So this seems fine.  However, alignment at scale 4 takes ~120 seconds with 4 cpus and with 1 cpu -- so we're seeing 0 speedup in this case.