This version saves results files in each scale directory for each layer.
The results files are saved in files named "single_alignment_out_#.json"
where "#" is the layer number being aligned. The files are stored in the
"scale_$" directory where "$" is the scale number.

Hopefully, this code won't be needed because alternate communication
methods (pipes) will work. This version was only being used as a back up
in case the pipe approach was not able to be sped up. Recent tests show
that there is great hope in that alternate approach so this code may
not be kept in the project.