Added Clear Role to allow for importing cropped images repeatedly.
The general purpose role capability still exists in AlignEM. So new
roles can be added at any time. However, the previous version had
no ability to clear the images from a role, so there was no way to
repeatedly crop and view images again and again. This commit adds
that capability to the tool.