Added an option to suppress display of skipped images.
This commit adds a change to the scroll handler to keep scrolling
(in the current direction) to find either the next unskipped image
or the end of the stack (which will show that image even if it is
skipped).

Note that a better implementation should reverse itself when it
finds the end of the stack but still no visible images (although
that could also be confusing since scrolling in one direction
might actually show an image from the opposite direction). This
version is being saved because it works in cases where images are
skipped in the "interior" of the stack. That's generally the most
useful situation. A further enhancement may be added to reverse
direction at the end of the stack (only once or it may be infinite).