Re-added the skip X drawing from earlier.
The breaking of the "skip" behavior appears to have happened earlier.

It will require further investigation to fix the skipping. At this
point, the skip flag appears to be valid and is properly saved in the
JSON project ... as long as the scroll wheel doesn't move from the
current location. As soon as it moves, the skip values are all set False.