BIG FIX for SKIPs: Use the clicked callback rather than stateChanged.
The QCheckBox widget provides a very nice stateChanged callback that
is demonstrated in almost every example on the internet. After all,
it is the "stateChanged" feature that makes a QCheckBox a QCheckBox.
Otherwise it would just be a regular button (clicked and released).

However, there's a catch with the "stateChanged" feature that's not
always obvious. The "stateChanged" event is generated regardless of
whether the change came from an actual mouse click or from a program
simply setting the value of the QCheckBox to "checked" or "unchecked".

This feature isn't usually a problem because most user interfaces
don't change the values of settings for the user. However, that's
exactly what we do when we scroll from layer to layer. Our program
must change those checkboxes as the layers are changed to reflect
the value of (for example) the skipped flag on that layer.

So every time we change a QCheckBox (due to a layer change or a
scale change), the user interface is calling our "stateChanged"
callback. This led to the implementation of an "ignore" flag that
was set prior to the program changing a value and cleared after
the value had been changed. But this might be subject to race
conditions since the timing isn't guaranteed. Furthermore, it's
messy to have to set and check these ignore flags for all of the
checkboxes we might use.

The solution appears to be to ignore the stateChanged signal and
just connect to the "clicked" signal. Testing confirms that the
"clicked" signal is ONLY generated when an actual mouse click
changes the checkbox. That's what we need.

This commit makes one of the first changes to use the "clicked"
signal instead of the "stateChanged" signal. There may be more
subtle issue with this change in the future, but this is already
a big improvement.